Split models into seperate files
