+++ src/drc/datamodel/models/__init__.py
@@ -1,0 +1,8 @@
+from .bestandseel import BestandsDeel # noqa
+from .enkelvoudig_informatieobject import (
+    EnkelvoudigInformatieObjectCanonical,
+    EnkelvoudigInformatieObject
+) # noqa
+from .gebruiksrechten import Gebruiksrechten # noqa
+from .informatieobject import InformatieObject # noqa
+from .object_informatieobject import ObjectInformatieObject # noqa
